make xcode12 work

+++ ios/debugproxy/socket_mover.go
@@ -1,6 +1,7 @@
 package debugproxy
 
 import (
+	"fmt"
 	"log"
 	"os"
 )
@@ -9,9 +10,20 @@ const realSocketSuffix = ".real_socket"
 
 func MoveSock(socket string) (string, error) {
 	newLocation := socket + realSocketSuffix
+	if fileExists(newLocation) {
+		return "", fmt.Errorf("there is already a file named: %s please remove it or restore original usbmuxd before starting the proxy", newLocation)
+	}
 	log.Printf("Moving socket %s to %s", socket, newLocation)
 	err := os.Rename(socket, newLocation)
 	return newLocation, err
+}
+
+func fileExists(filename string) bool {
+	info, err := os.Stat(filename)
+	if os.IsNotExist(err) {
+		return false
+	}
+	return !info.IsDir()
 }
 
 func MoveBack(socket string) error {
